rename twitter workflow directory to directedTwitter
