+++ src/agents/workflows/directedTwitter/nodes/analyzeTrendNode.ts
